Fix shaders causing floating point exception

+++ src/example/gi_specular/voxel/world.frag
@@ -10,7 +10,7 @@ layout (set=1, binding = 2, r32ui) uniform uimage3D dest_image;
 void main()  {
   const ivec3 image_size = imageSize( dest_image );
   //const vec3 world_pos = vec3( ( input_position.xy + 1.0 )/ 2.0 * 256.0, input_position.z * 256.0 );
-  const ivec3 world_pos = ivec3( ( input_position.xyz ) * ( image_size.xyz - vec3( 1, 1, 1 ) ) );
+  const ivec3 world_pos = ivec3( ( input_position.xyz ) * ( image_size.xyz - ivec3( 1, 1, 1 ) ) );
   imageStore( dest_image, world_pos, uvec4( ( world_pos.x ) << 20 | ( world_pos.y << 10 ) | world_pos.z, 0, 0, 0 ) );
 }
